requirements.txt
APScheduler==3.10.4
Django==4.2.11
Pillow==9.3.0
django_apscheduler==0.6.2
django_cors-headers==4.3.1
djangorestframework==3.15.2
drf_spectacular==0.27.2
drf_spectacular_sidecar==2024.7.1
loguru==0.7.2
qbittorrent_api==2024.2.59
Requests==2.32.3
websockets==12.0
xmltodict==0.13.0
